Update dependency modules; split repo into submodules
This change updates dcrwallet to perform a module-aware build with the latest
versions of its dependencies.  It also splits the repo from being a single
module to a collection of multiple modules for package consumption by other
projects.

All dcrwallet modules used by other dcrwallet modules use 'replace' directives
in their go.mod files to reference the dependency using a relative file path.
This allows the repo to have split published modules, and to refer to
unpublished module versions before tagging, while still allowing local
development and testing using the most recent changes of any given imported
major version of a module.

The following modules and versions are provided by this change:

  github.com/decred/dcrwallet/chain@v1.0.0
  github.com/decred/dcrwallet/deployments@v1.0.0
  github.com/decred/dcrwallet/errors@v1.0.0
  github.com/decred/dcrwallet/internal/helpers@v1.0.0
  github.com/decred/dcrwallet/internal/zero@v1.0.0
  github.com/decred/dcrwallet/lru@v1.0.0
  github.com/decred/dcrwallet/p2p@v1.0.0
  github.com/decred/dcrwallet/pgpwordlist@v1.0.0
  github.com/decred/dcrwallet/spv@v1.0.0
  github.com/decred/dcrwallet/ticketbuyer@v1.0.0
  github.com/decred/dcrwallet/ticketbuyer/v2@v2.0.0
  github.com/decred/dcrwallet/validate@v1.0.0
  github.com/decred/dcrwallet/wallet@v1.0.0
  github.com/decred/dcrwallet/walletseed@v1.0.0

A module in the project root also exists to provide reproducible builds, but
packages in this module are not intended for consumption as libraries for other
projects.

The repo must be tagged following this change to publish these modules and
versions to outside projects.
