+++ deployments/go.mod
@@ -1,0 +1,6 @@
+module github.com/decred/dcrwallet/deployments
+
+require (
+	github.com/decred/dcrd/chaincfg v1.1.1
+	github.com/decred/dcrd/wire v1.1.0
+)
