+++ internal/helpers/go.mod
@@ -1,0 +1,9 @@
+module github.com/decred/dcrwallet/internal/helpers
+
+require (
+	github.com/decred/dcrd/dcrec v0.0.0-20180809193022-9536f0c88fa8 // indirect
+	github.com/decred/dcrd/dcrec/edwards v0.0.0-20180809193022-9536f0c88fa8 // indirect
+	github.com/decred/dcrd/dcrutil v1.1.1
+	github.com/decred/dcrd/wire v1.1.0
+	golang.org/x/crypto v0.0.0-20180808211826-de0752318171 // indirect
+)
