+++ internal/zero/benchmark_test.go
@@ -1,14 +1,12 @@
 // Copyright (c) 2015 The btcsuite developers
-// Copyright (c) 2015 The Decred developers
+// Copyright (c) 2015-2018 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package zero_test
+package zero
 
 import (
 	"testing"
-
-	. "github.com/decred/dcrwallet/internal/zero"
 )
 
 var (
