+++ internal/zero/go.mod
@@ -1,0 +1,5 @@
+module github.com/decred/dcrwallet/internal/zero
+
+require github.com/decred/dcrwallet/errors v1.0.0
+
+replace github.com/decred/dcrwallet/errors => ../../errors
