+++ internal/zero/zero_test.go
@@ -1,9 +1,9 @@
 // Copyright (c) 2015 The btcsuite developers
-// Copyright (c) 2015 The Decred developers
+// Copyright (c) 2015-2018 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
-package zero_test
+package zero
 
 import (
 	"math/big"
@@ -11,7 +11,6 @@ import (
 	"testing"
 
 	"github.com/decred/dcrwallet/errors"
-	. "github.com/decred/dcrwallet/internal/zero"
 )
 
 func makeOneBytes(n int) []byte {
